+++ kernel/trace/trace_kprobe.c
@@ -1421,7 +1421,7 @@ process_fetch_insn(struct fetch_insn *code, void *rec, void *edata,
 		   void *dest, void *base)
 {
 	struct pt_regs *regs = rec;
-	unsigned long val;
+	uintptr_t val;
 	int ret;
 
 retry:
